+++ internal/tests/features/register_proxy_test.go
@@ -20,7 +20,7 @@ func Test_Register_generated_proxy_type(t *testing.T) {
 	_ = registry.Register(newMethodCallInterceptor(collector), types.LifetimeSingleton)
 	_ = registry.Register(NewGreeterProxyImpl, types.LifetimeTransient)
 	_ = registry.Register(newGreeter, types.LifetimeTransient)
-	_ = features.RegisterList[features.MethodInterceptor](ctx, registry)
+	_ = features.RegisterList[features.MethodInterceptor](registry)
 
 	resolver := resolving.NewResolver(registry)
 	resolverContext := resolving.NewScopedContext(ctx)